Fix motor auto detection
The automatic detection of available motors was based on rotor_data having a
nul timestamp. This is fragile, and no longer works since the timestamps are
now used to detected fresh data for loggging.

Instead, use a dedicated array which stores the autodection status. The bits
are true until an explicit enable/disable motor is done, which disables the
autoconfiguration.