Do not go to 'nodata' state when poll(2) is interrupted in mk_comm_poll
It can happen that poll(2) in mk_wait_msg() is interrupted. This does not
imply that no data has been received, especially since the 'nodata' state
resets motor data and would erroneously stop the servo() activity due to the
'motor spinning' state check.

Spotted with the help and patience of Lev Smolentsev, thanks!